Add cron to update stable branch (#20)
Using submodules means that `develop` has to be manually updated. But we
can create a cron that will ensure that `feature/stable-nightly` has the
most recent versions of JEDI repositories that build and tests pass.

To do this, the default branches had to be updated/added for the
repositories. This is because the public JCSDA repos have `master`
branches as default but we usually need `develop`. Running `git
submodule update` on the public repos would have resulted in grabbing
very old commits rather than recent ones.

Note: this script will not run properly until develop is updated with
those submodule changes in `.gitmodules`. To test, one can add `git
checkout feature/update-stable` in the `stable_driver.sh` script right
after `git checkout develop` but this was tested manually yesterday on
Orion. See:
https://github.com/NOAA-EMC/RDASApp/tree/feature/stable-nightly