Add EVA tools for validation (#42)
The purpose of this PR is to add the EVA tools that exist in
GDASApp/ush/eva into RDASApp/ush/eva. In addition, additions were made
to the setup_experiment script for a user to select if they would like
to set up the EVA tools as well. Those scripts are then copied to the
experiment directory in addition to a sample helper script to help users
get started running EVA. Some of the non-EVA validation/plotting tools
were also refactored which can and should be used in combination with
the EVA tools. The wiki will be updated as needed.

This PR follows the following wiki: [Running FV3‐JEDI Hofx
Validation](https://github.com/NOAA-EMC/RDASApp/wiki/Running-FV3%E2%80%90JEDI-Hofx-Validation)
specifically the EVA validation tools near the bottom