Add LETKF/GETKF yamls for new MPAS-JEDI test case (#125)
This PR adds yaml files for LETKF and GETKF for the new 2024052700
MPAS-JEDI test case. The original yaml file for this case has been
renamed to `rrfs-test/testinput/rrfs_mpasjedi_2024052700_Ens3Dvar.yaml`
to be consistent with the new files and our older test cases. The run
scripts for this new case are also updated with a line that can be
uncommented to run these EnKF yamls instead of EnVar.