Update offline domain check script (#153)
This program determines if observations are in/outside of a convex hull compared via a lat/lon grid file. A convex hull is the smallest convex shape (or polygon) that can enclose a set of points in a plane (or in higher dimensions). Imagine stretching a rubber band around the outermost points in a set; the shape that the rubber band forms is the convex hull. So, if there are any concave points between vertices, then there would be a whitespace between the red and blue boxes. I've shrunk the convex hull such that there wouldn't be such whitespace which, of course, in turn, means that it is going to be not an exact match of the domain grid (e.g., near corners). This can be tuned via the "hull_shrink_factor".