Add ctests for running GETKF in split mode and move reference files into repo  (#175)
This PR adds ctests and yaml files for running GETKF in split observer
and solver mode. Running GETKF in this form is currently the most
efficient option since the observer can be run using the RoundRobin
distribution (see #122). Reference files are also added for these two
new ctests and are added into the repo (see below).