Add utility for updating ctest input files during development (#202)
This PR adds a simple shell script based on `rrfs-test/CMakeLists.txt`
that allows developers to update the RRFS ctests in RDASApp. The goal
here is to be able to run this quick script if making any changes to the
yaml files, test reference files, or input data for the ctests.
Previously, the only way to do this was to rebuild all of RDASApp or
make temporary changes to `build.sh`

Please see
https://github.com/NOAA-EMC/RDASApp/pull/184#discussion_r1797962906 for
additional motivation.

Note: this script cannot update the ctest configuration (number of
tests, test names, or MPI configuration) since those options are handled
by `ecbuild` during the cmake step.