A basic sample Yaml file for the new integrated operator for surface observations with consistent adjustments (#152)
A new UFO operator "SfcCorrected" is proposed to integrate surface
temperature, humidity, and wind observations (and pressure later) with
various adjustments. The main motivation is to ensure consistent
empirical adjustments to simulations (HofX) and optimal assimilation of
the observations. It would also enable simpler configurations of the
observations in a yaml file.
First, a few terrain height adjustments for T2m are included. The
options of "WRFDA" and "UKMO" are consistent with the air temperature
assumptions between model surface and observation station height used in
the existing surface pressure operator "SfcPCorrected". The "GSL" option
is intended for the experimental Rapid Refresh Forecast System (RRFS).
The existing pressure operator could be blended into this new operator
later.
For specific humidity and wind, empirical height adjustments could be
evaluated and added later, like the UKMO options in the obsfunctions/.
Wind is simulated from model surface level and scaled by a ratio of
model diagnosed 10m wind over model surface wind, which is calculated
online in JEDI.

This basic yaml could serve as a template to add options for specific
obs types and purpose.

Dependent:
https://github.com/JCSDA-internal/ufo/pull/3440

---------

Co-authored-by: Guoqing Ge <guoqing.ge@noaa.gov>
Co-authored-by: Samuel Degelia <samuel.degelia@noaa.gov>