remove no-longer-needed `diagnostics.departures` and rename jedi diagnostic files as jdiag* (#241)
As discussed in issue #240 , we no longer need to specify `diagnostics.departures` in YAML files and JEDI will always output
`ombg, oman, innov1` for two outer loop minimization
Also, the JEDI diagnostic file contains much more information than `hofx`. It is better to rename the JEDI diagnostic files as `jdiag*`, for example `jdiag_adpupa.nc4` to be consistent with the reality (the word `jdiag` has been used for a long time in the python plottingscripts).