Phase 1 validation: adpupa (132/232) yamls (#323)
Adding phase 1 adpupa (132/232) yamls. There were no obs in the ctest
case to fully validate. This isn't a surprise since these ob types were
marked to be pretty sparse in the observation spreadsheet. These yamls
were developed following the (120/220) yamls exactly with the exception
of the ObsErrors and gross error checks for winds. The point of this PR
is to establish these yamls in RDASApp. This won't affect cycling DA
results because we will only use configurations that have passed phase 3
validation.