Add wrapper script for running ctests on WCOSS2 (#377)
This PR adds a wrapper script `wrapper_ctest_pbs.sh` for running ctests
on WCOSS2. The ctests here can not be submitted using `mpirun` through
the command line as is normally done with the `ctest` command. Running
`mpirun` for PBS on WCOSS2 expects different input than what is defined
in our `CMakeLists.txt` file including a host list that is not possible
to assign automatically. Instead we introduce this wrapper script that
creates and submits a job card for each ctest.