Update path to `RDAS_DATA` on WCOSS2 (#422)
With https://github.com/NOAA-EMC/rrfs-workflow/pull/893, we updated the
paths to `FIX_RRFS2` and `RRFS2_RETRO_DATA` now that we have settled on
a permanent location to store those data on WCOSS2. This PR
correspondingly updates the path to `RDAS_DATA` that is now saved in the
same location.