Add python virtual environment for running offline domain checks on WCOSS2 (#464)
Some python libraries are not available in the spack-stack 1.9 version
loaded on WCOSS2, including `matplotlib` and `cartopy`. These libraries
are needed to run the offline domain check codes. Here, we add a new
python virtual environment to the RDAS module file on WCOSS2 that
includes these python libraries. This `rdas-mpl-311` virtual environment
is built on top of the python version loaded in spack-stack so there
should be no conflicts.