Hotfix for determining spack-stack version on Jet (#476)
This hotfix changes the logic used to determine which spack-stack version is loaded to copy in the relevant gsibec workaround codes. The previous version did not work with spack-stack 1.9 on Jet which was causing issues with #419 and #475.
A fix is also included to apply the new gsibec parameters to all of the yamls now that we have added more ctests.
After this fix, I found that all of the ctests in #419 are passing when updating to spack-stack 1.9 on Jet.

This should not affect anything on Jet prior to updating to spack-stack 1.9. But I will go ahead and run the CI tests to confirm that this does not break anything on other machines.

**Description of Current Behavior:**
1. When upgrading trying to use spack-stack 1.9 on Jet, the logic in `build.sh` to copy in the gsibec specific code fails.
2. The incorrect gsibec files are thus used, leading to mismatches between the Atlas and GSI grids.