[Hotfix] Remove the diagnostic output files from MGBF  (#479)
The current workaround MGBF code outputs a number of diagnostic files
for each PE. While useful for development and debugging, these files can
also make it hard to parse through the run directories and are not
needed in rrfs-workflow. This PR removes those output files to leave a
cleaner run directory.