Fix ctest failure for getkf solver on Hera by disabling posterior observer (#486)
This PR resolves a common ctest failure for
`rrfs_fv3jedi_2024052700_getkf_solver` on Hera (see #480). It was found
that the test was stalling when writing out the posterior observer to
the jdiag files. I am still not sure why this stall occurred. But until
then, we can resolve the failure simply by disabling the posterior
observer.