Add workaround fix for `ObsErrorFactorPressureCheck`  (#491)
@spanNOAA found the
issue with the `ObsErrorFactorPressureCheck` filter that led to a
segmentation fault when run for `adpupa_winds` observations on Hercules.
It was caused by a index-searching bug in the vertical interpolation
routine. His workaround is also included here in RDASApp so that we can resolve
the ctest failures and pass the fix into the `[dev-sci]` branch of
rrfs-workflow.