Update convinfo vs j2 comparison tool and add offline duplicate tagging tool (#499)
This PR updates a tool and adds yet another offline tool
- Updated tool (`check_j2_convinfo.py`): Updates makes it work for
satwnd j2 file. Simply run this tool without any arguments and it checks
the convinfo settings vs what is actually prescribed in each j2
template.
- New tool (`offline_duplicate_tagger.py`): Takes two ioda files from
different cycles (e.g., 00Z and 01Z both ioda_adpupa.nc) and find
duplicated (from cycle-to-cycle). This is possible because our time
window is +/-1.5h and the exact same obs are possibly assimilated more
more than one cycle while treating them as independent obs. The goal at
the moment is to just run this in the workflow as diagnostic only to get
a better understanding of how many obs are potentially duplicate (from
cycle-to-cycle). This tool simply adds a new metadata field and mark it
as either 0=new or 1=usedPreviously or -1=default. In the future, we
could use this metadata field to filter out those duplicate obs.