Set observation error in yaml for `mrms_refl` instead of reading it from IODA file  (#506)
This PR is a small follow up to #504 that defines the observation error
for reflectivity directly in the yaml file rather than reading it from
the IODA file. This should not affect any results since the observation
errors in the IODA file were already set to 5 dBZ. But the change makes
the reflectivity configuration consistent with how observation errors
are handled in the rest of the yamls.

We also update `validate_ob_filters.py` to skip enforcing the presence
of a where clause for a small set of observation spaces that do not
contain an `ObsType` group in their IODA files. Without this change,
those observation spaces would continuously fail validation due to
missing where statements in their filters.