Add JCB template files for running Hybrid 3DEnVar for reflectivity DA  (#509)
This is a small PR to add two extra JCB template files (one for MGBF and
one for BUMP) to run the Hybrid 3DEnVar when assimilating reflectivity
observations. These are not used directly in RDASApp since our radar DA
ctest uses pure ensemble 3DEnVar. But these templates are needed for
rrfs-workflow where we plan to assimilate reflectivity obs using the
hybrid.