Fixes for moisture linear variable change in fv3-jedi and introduction of `qoption=2` in `gsibec` (#510)
This PR introduces a workaround to resolve a bug in fv3-jedi with the
linear variable change for moisture. It also modifies the gsibec
workarounds to allow fv3-jedi to now work with `qoption=2` (normalized
RH). Thanks to @Masanori-NOAA for his great work in solving this
problem.

In addition to the updated workarounds, we also updated the JCB
templates to include new gsibec parameters. The test references are also
updated given that the new workarounds change the results for any test
using gsibec.